macro supports lifetimes
